HDDS-11999. Remove ozone repair ldb command and move its subcommands to ozone repair om
Remove  command and move its subcommands to

Corrected TestOzoneRepairShell
